Refactor memory module into routes and core
