+++ core/memoryOperations.js
@@ -1,0 +1,720 @@
+const fs = require('fs');
+const path = require('path');
+const github = require('../githubClient');
+const tokenStore = require('../tokenStore');
+const memoryConfig = require('../memoryConfig');
+const mdEditor = require('../markdownEditor');
+const fileEditor = require('../markdownFileEditor');
+const validator = require('../markdownValidator');
+const {
+  ensureDir,
+  deepMerge,
+  normalizeMemoryPath,
+} = require('../utils/fileUtils');
+const {
+  getRepoInfo,
+  categorizeMemoryFile,
+  logDebug,
+} = require('../utils/memoryHelpers');
+
+const contextFilename = path.join(__dirname, '..', 'memory', 'context.md');
+const planFilename = path.join(__dirname, '..', 'memory', 'plan.md');
+const indexFilename = path.join(__dirname, '..', 'memory', 'index.json');
+
+let planCache = null;
+
+function parsePlanMarkdown(md) {
+  const plan = { completedLessons: [], requestedClarifications: [], nextLesson: '' };
+  const lines = md.split(/\r?\n/);
+  let section = '';
+  for (const line of lines) {
+    if (/^##\s+Completed Lessons/i.test(line)) { section = 'completed'; continue; }
+    if (/^##\s+Requested Clarifications/i.test(line)) { section = 'clar'; continue; }
+    if (/^##\s+Progress/i.test(line)) { section = 'progress'; continue; }
+    if (/^-\s+/.test(line) && section === 'completed') {
+      plan.completedLessons.push(line.replace(/^-\s+/, '').trim());
+    } else if (/^-\s+/.test(line) && section === 'clar') {
+      plan.requestedClarifications.push(line.replace(/^-\s+/, '').trim());
+    } else if (section === 'progress') {
+      const m = line.match(/Next lesson:\s*(.*)/i);
+      if (m) plan.nextLesson = m[1].trim();
+    }
+  }
+  return plan;
+}
+
+function planToMarkdown(plan) {
+  const completed = (plan.completedLessons || []).map(t => `- ${t}`).join('\n');
+  const clar = (plan.requestedClarifications || []).map(t => `- ${t}`).join('\n');
+  const total = plan.totalLessons || (plan.completedLessons || []).length;
+  return (
+    `# Learning Plan\n\n` +
+    `## Completed Topics\n${completed}\n\n` +
+    `## Clarified or Expanded Topics\n${clar}\n\n` +
+    `## Progress: ${plan.completedLessons.length} / ${total} lessons complete` +
+    (plan.nextLesson ? `\nNext lesson: ${plan.nextLesson}` : '') +
+    `\n`
+  );
+}
+
+function getLessonEntries() {
+  const raw = sanitizeIndex(readIndexSafe());
+  return raw.filter(e => e.type === 'lesson');
+}
+
+function updatePlanFromIndex(plan) {
+  const lessons = getLessonEntries();
+  const titles = new Set();
+  plan.completedLessons = [];
+  lessons
+    .sort((a, b) => {
+      if (a.lastModified && b.lastModified) {
+        return new Date(a.lastModified) - new Date(b.lastModified);
+      }
+      return (a.path || '').localeCompare(b.path || '');
+    })
+    .forEach(l => {
+      const title = l.title || path.basename(l.path || '');
+      if (!titles.has(title)) {
+        titles.add(title);
+        plan.completedLessons.push(title);
+      }
+    });
+  plan.totalLessons = lessons.length;
+  return plan;
+}
+
+function ensureContext() {
+  if (!fs.existsSync(contextFilename)) {
+    ensureDir(contextFilename);
+    writeFileSafe(contextFilename, '# Context\n');
+    rebuildIndex().catch(e => console.error('[ensureContext] rebuild error', e.message));
+  }
+}
+
+function parsePlanFile(md) {
+  const lines = md.split(/\r?\n/);
+  let section = null;
+  const result = {
+    lines,
+    done: [],
+    upcoming: [],
+    doneStart: -1,
+    doneEnd: -1,
+    upcomingStart: -1,
+    upcomingEnd: -1
+  };
+
+  for (let i = 0; i < lines.length; i++) {
+    const line = lines[i];
+    const header = line.match(/^##\s*(.+)/);
+    if (header) {
+      const name = header[1].toLowerCase();
+      if (section === 'done') result.doneEnd = i;
+      if (section === 'upcoming') result.upcomingEnd = i;
+      if (name.startsWith('completed')) {
+        section = 'done';
+        result.doneStart = i + 1;
+      } else if (name.startsWith('upcoming')) {
+        section = 'upcoming';
+        result.upcomingStart = i + 1;
+      } else {
+        section = null;
+      }
+      continue;
+    }
+
+    if (/^[-*]\s+/.test(line)) {
+      const item = line.replace(/^[-*]\s+/, '').replace(/^✅\s*/, '').trim();
+      if (section === 'done') result.done.push(item);
+      if (section === 'upcoming') result.upcoming.push(item);
+    }
+  }
+
+  if (section === 'done') result.doneEnd = lines.length;
+  if (section === 'upcoming') result.upcomingEnd = lines.length;
+  return result;
+}
+
+function buildPlanFile(parsed, done, upcoming) {
+  const lines = parsed.lines.slice();
+
+  const doneLines = done.map(t => `- ✅ ${t}`);
+  if (parsed.doneStart >= 0) {
+    lines.splice(parsed.doneStart, parsed.doneEnd - parsed.doneStart, ...doneLines);
+  } else {
+    lines.push('## Completed Lessons', ...doneLines, '');
+  }
+
+  const upcomingLines = upcoming.map(t => `- ${t}`);
+  if (parsed.upcomingStart >= 0) {
+    lines.splice(parsed.upcomingStart, parsed.upcomingEnd - parsed.upcomingStart, ...upcomingLines);
+  } else {
+    lines.push('## Upcoming Lessons', ...upcomingLines, '');
+  }
+
+  return lines.join('\n');
+}
+
+async function updatePlan({ token, repo, updateFn, userId } = {}) {
+  const relPath = 'memory/plan.md';
+  const absPath = path.join(path.join(__dirname, '..'), relPath);
+
+  const { repo: finalRepo, token: finalToken } = getRepoInfo(relPath, userId, repo, token);
+
+  let md = '';
+  if (finalRepo && finalToken) {
+    try {
+      md = await github.readFile(finalToken, finalRepo, relPath);
+    } catch (e) {
+      logDebug('[updatePlan] no remote plan', e.message);
+    }
+  }
+  if (!md && fs.existsSync(absPath)) {
+    md = fs.readFileSync(absPath, 'utf-8');
+  }
+  if (!md) {
+    md = '# Learning Plan\n\n## Completed Lessons\n\n## Upcoming Lessons\n';
+  }
+
+  const parsed = parsePlanFile(md);
+  const current = { done: parsed.done.slice(), upcoming: parsed.upcoming.slice() };
+  const updated = updateFn ? updateFn(current) : current;
+
+  const newMd = buildPlanFile(parsed, updated.done || [], updated.upcoming || []);
+
+  writeFileSafe(absPath, newMd);
+  if (finalRepo && finalToken) {
+    await githubWriteFileSafe(finalToken, finalRepo, relPath, newMd, 'update plan.md');
+  }
+
+  console.log('[updatePlan] ✅ plan.md updated');
+  return updated;
+}
+
+function loadPlan() {
+  ensureDir(planFilename);
+  const existed = fs.existsSync(planFilename);
+  let plan;
+  if (existed) {
+    try {
+      const content = fs.readFileSync(planFilename, 'utf-8');
+      plan = parsePlanMarkdown(content);
+    } catch {
+      plan = { completedLessons: [], requestedClarifications: [], nextLesson: '' };
+    }
+  } else {
+    plan = { completedLessons: [], requestedClarifications: [], nextLesson: '' };
+  }
+
+  plan = updatePlanFromIndex(plan);
+  writeFileSafe(planFilename, planToMarkdown(plan));
+  planCache = plan;
+  if (!existed) {
+    rebuildIndex().catch(e => console.error('[loadPlan] rebuild error', e.message));
+  }
+}
+
+async function savePlan(repo, token) {
+  if (!planCache) loadPlan();
+  const md = planToMarkdown(planCache);
+  writeFileSafe(planFilename, md);
+  if (repo && token) {
+    try {
+      await githubWriteFileSafe(token, repo, path.relative(path.join(__dirname, '..'), planFilename), md, 'update plan');
+    } catch (e) {
+      console.error('GitHub write plan error', e.message);
+    }
+  }
+  await rebuildIndex(repo, token);
+}
+
+function writeFileSafe(filePath, data, force = false) {
+  try {
+    ensureDir(filePath);
+    if (filePath.toLowerCase().endsWith('.md')) {
+      const check = validator.validateMarkdownSyntax(data, filePath);
+      if (!check.valid) {
+        const backup = mdEditor.createBackup(filePath);
+        console.error(
+          `[writeFileSafe] ${check.message} at line ${check.line} in '${path.basename(filePath)}'`
+        );
+        if (backup) {
+          console.error(`[writeFileSafe] You can restore from: ${backup}`);
+        }
+        if (!force) return;
+      }
+      mdEditor.createBackup(filePath);
+    }
+    fs.writeFileSync(filePath, data, 'utf-8');
+    logDebug('[writeFileSafe] wrote', filePath);
+  } catch (e) {
+    console.error(`[writeFileSafe] Error writing ${filePath}`, e.message);
+    throw e;
+  }
+}
+
+async function githubWriteFileSafe(token, repo, relPath, data, message, attempts = 2) {
+  for (let i = 1; i <= attempts; i++) {
+    try {
+      ensureDir(path.join(path.dirname(__dirname), path.dirname(relPath)));
+      await github.writeFile(token, repo, relPath, data, message);
+      logDebug('[githubWriteFileSafe] pushed', relPath);
+      return;
+    } catch (e) {
+      console.error(`[githubWriteFileSafe] attempt ${i} failed for ${relPath}`, e.message);
+      if (i === attempts) throw e;
+    }
+  }
+}
+
+async function updateOrInsertJsonEntry(filePath, newData, matchKey, repo, token) {
+  ensureDir(filePath);
+  const relPath = path.relative(path.join(__dirname, '..'), filePath);
+  let existing = Array.isArray(newData) ? [] : {};
+
+  if (repo && token) {
+    try {
+      const remote = await github.readFile(token, repo, relPath);
+      existing = JSON.parse(remote);
+    } catch {
+      // ignore missing remote file
+    }
+  }
+
+  if (fs.existsSync(filePath)) {
+    try {
+      const local = fs.readFileSync(filePath, 'utf-8');
+      existing = deepMerge(existing, JSON.parse(local), matchKey);
+    } catch {
+      // ignore parse errors
+    }
+  }
+
+  const updated = deepMerge(existing, newData, matchKey);
+  writeFileSafe(filePath, JSON.stringify(updated, null, 2));
+
+  if (repo && token) {
+    try {
+      await githubWriteFileSafe(
+        token,
+        repo,
+        relPath,
+        JSON.stringify(updated, null, 2),
+        `update ${path.basename(filePath)}`
+      );
+    } catch (e) {
+      console.error('GitHub write error', e.message);
+    }
+  }
+
+  return updated;
+}
+
+async function fileExistsInRepo(repo, token, relPath) {
+  if (!repo || !token) return true;
+  try {
+    await github.readFile(token, repo, relPath);
+    return true;
+  } catch (e) {
+    if (e.response?.status === 404) return false;
+    console.error('[fileExistsInRepo] GitHub check failed', e.message);
+    return true;
+  }
+}
+
+function deduplicateEntries(entries) {
+  const map = new Map();
+  const result = [];
+  entries.forEach(e => {
+    const base = path.basename(e.path).toLowerCase().replace(/\.[^.]+$/, '');
+    const key = (e.title || '').toLowerCase() || base.replace(/[-_].*/, '');
+    if (map.has(key)) {
+      console.warn(`[deduplicateEntries] duplicate ${e.path} ignored; matches ${map.get(key).path}`);
+    } else {
+      map.set(key, e);
+      result.push(e);
+    }
+  });
+  return result;
+}
+
+const EXCLUDED = new Set([
+  'memory/index.json',
+  'memory/plan.md',
+  'memory/context.md',
+  'memory/test.md',
+  'memory/test.txt',
+]);
+
+function readIndexSafe() {
+  if (!fs.existsSync(indexFilename)) return [];
+  try {
+    const content = fs.readFileSync(indexFilename, 'utf-8');
+    const data = JSON.parse(content);
+    return Array.isArray(data) ? data : [];
+  } catch {
+    return [];
+  }
+}
+
+function sanitizeIndex(entries) {
+  const map = new Map();
+  entries.forEach(e => {
+    if (!e || !e.path) return;
+    const normalized = normalizeMemoryPath(e.path);
+    if (EXCLUDED.has(normalized)) return;
+    const abs = path.join(path.join(__dirname, '..'), normalized);
+    if (!fs.existsSync(abs)) {
+      console.warn(`[sanitizeIndex] missing file ${normalized}, keeping entry`);
+    }
+    const existing = map.get(normalized);
+    if (!existing || new Date(e.lastModified || 0) > new Date(existing.lastModified || 0)) {
+      map.set(normalized, { ...existing, ...e, path: normalized });
+    }
+  });
+  return Array.from(map.values());
+}
+
+function extractMeta(fullPath) {
+  const stats = fs.statSync(fullPath);
+  const result = { lastModified: stats.mtime.toISOString() };
+  if (fullPath.endsWith('.md')) {
+    try {
+      const lines = fs.readFileSync(fullPath, 'utf-8').split(/\r?\n/);
+      const titleLine = lines.find(l => l.trim());
+      if (titleLine && titleLine.startsWith('#')) {
+        result.title = titleLine.replace(/^#+\s*/, '');
+      }
+      if (lines.length > 1) {
+        result.description = lines.slice(1, 3).join(' ').slice(0, 100);
+      }
+    } catch {
+      // ignore parsing errors
+    }
+  }
+  return result;
+}
+
+function loadIndex() {
+  if (!fs.existsSync(indexFilename)) {
+    console.warn('[loadIndex] index.json not found - creating new');
+    ensureDir(indexFilename);
+    writeFileSafe(indexFilename, '[]');
+    return [];
+  }
+
+  try {
+    const parsed = JSON.parse(fs.readFileSync(indexFilename, 'utf-8'));
+    return Array.isArray(parsed) ? parsed : [];
+  } catch (e) {
+    console.warn('[loadIndex] failed to parse index.json, resetting', e.message);
+    writeFileSafe(indexFilename, '[]');
+    return [];
+  }
+}
+
+function saveIndex(data) {
+  ensureDir(indexFilename);
+  writeFileSafe(indexFilename, JSON.stringify(data, null, 2));
+}
+
+async function updateIndexFile(entry, repo, token, userId) {
+  const data = await fetchIndex(repo, token);
+  const idx = data.findIndex(i => i.path === entry.path);
+  if (idx >= 0) {
+    data[idx] = { ...data[idx], ...entry };
+    console.log('[updateIndexFile] updated', entry.path);
+  } else {
+    data.push(entry);
+    console.log('[updateIndexFile] added', entry.path);
+  }
+  const clean = sanitizeIndex(deduplicateEntries(data));
+  await persistIndex(clean, repo, token, userId);
+  return clean;
+}
+
+async function updateIndexFromPath(relPath, repo, token, userId) {
+  const fullPath = path.join(path.join(__dirname, '..'), relPath);
+  if (!fs.existsSync(fullPath)) return;
+  const meta = extractMeta(fullPath);
+  const entry = {
+    path: relPath,
+    type: categorizeMemoryFile(path.basename(relPath)),
+    ...meta,
+  };
+
+  await updateIndexFile(entry, repo, token, userId);
+}
+
+async function scanMemoryFolderRecursively(repo, token, basePath = 'memory') {
+  const files = [];
+
+  function walk(current) {
+    const items = fs.readdirSync(current, { withFileTypes: true });
+    items.forEach(item => {
+      const abs = path.join(current, item.name);
+      if (item.isDirectory()) {
+        walk(abs);
+      } else if (item.isFile()) {
+        const rel = path.relative(path.join(__dirname, '..'), abs).replace(/\\/g, '/');
+        if (rel.endsWith('index.json')) {
+          logDebug('[scan] skipped', rel, 'index file');
+          return;
+        }
+        if (/\.(md|txt|json|js|ts|jsx|tsx|html|css|png|jpe?g|svg|gif|py|java|c|cpp|csv)$/i.test(item.name)) {
+          files.push(rel);
+          logDebug('[scan] file', rel);
+        } else {
+          logDebug('[scan] skipped', rel, 'unsupported extension');
+        }
+      }
+    });
+  }
+
+  const rootPath = path.join(path.join(__dirname, '..'), basePath);
+  if (fs.existsSync(rootPath)) walk(rootPath);
+
+  const verified = [];
+  for (const rel of files) {
+    // eslint-disable-next-line no-await-in-loop
+    if (await fileExistsInRepo(repo, token, rel)) {
+      verified.push(rel);
+    } else {
+      console.warn(`[scanMemoryFolderRecursively] ${rel} not found in repo`);
+    }
+  }
+  return verified;
+}
+
+async function fetchIndex(repo, token) {
+  const indexRel = path.relative(path.join(__dirname, '..'), indexFilename);
+  let localData = [];
+  let remoteData = [];
+
+  if (fs.existsSync(indexFilename)) {
+    try {
+      const parsed = JSON.parse(fs.readFileSync(indexFilename, 'utf-8'));
+      if (Array.isArray(parsed)) localData = parsed;
+    } catch (e) {
+      console.warn('[fetchIndex] local read error', e.message);
+    }
+  }
+
+  if (repo && token) {
+    try {
+      const remote = await github.readFile(token, repo, indexRel);
+      const parsedRemote = JSON.parse(remote);
+      if (Array.isArray(parsedRemote)) remoteData = parsedRemote;
+    } catch (e) {
+      if (e.response?.status !== 404) {
+        console.error('[fetchIndex] GitHub read error', e.message);
+      }
+    }
+  }
+
+  const map = new Map();
+  [...localData, ...remoteData].forEach(entry => {
+    if (entry && entry.path) {
+      map.set(entry.path, { ...map.get(entry.path), ...entry });
+    }
+  });
+
+  return Array.from(map.values());
+}
+
+async function persistIndex(data, repo, token, userId) {
+  const clean = sanitizeIndex(data);
+  ensureDir(indexFilename);
+  try {
+    writeFileSafe(indexFilename, JSON.stringify(clean, null, 2));
+    console.log('[persistIndex] local index saved');
+  } catch (e) {
+    console.error('[persistIndex] local write error', e.message);
+  }
+
+  const finalRepo = repo || (userId ? memoryConfig.getRepoUrl(userId) : memoryConfig.getRepoUrl());
+  const finalToken = token || (userId ? tokenStore.getToken(userId) : null);
+
+  if (finalRepo && finalToken) {
+    try {
+      await githubWriteFileSafe(
+        finalToken,
+        finalRepo,
+        path.relative(path.join(__dirname, '..'), indexFilename),
+        JSON.stringify(clean, null, 2),
+        'update index.json'
+      );
+      console.log('[persistIndex] pushed index to GitHub');
+    } catch (e) {
+      console.error('[persistIndex] GitHub write error', e.message);
+    }
+  }
+}
+
+async function updateIndexEntry(repo, token, { path: filePath, type, title, description, lastModified }, userId) {
+  if (!filePath) return null;
+
+  const normalized = normalizeMemoryPath(filePath);
+  const indexData = await fetchIndex(repo, token);
+  const idx = indexData.findIndex(e => path.posix.normalize(e.path) === normalized);
+
+  const entry = {
+    path: normalized,
+    type: type || categorizeMemoryFile(path.basename(normalized)),
+  };
+  if (title !== undefined) entry.title = title;
+  if (description !== undefined) entry.description = description;
+  entry.lastModified = lastModified || new Date().toISOString();
+
+  let op = 'skipped';
+  if (idx >= 0) {
+    const existing = indexData[idx];
+    const metaChanged =
+      existing.type !== entry.type ||
+      existing.title !== entry.title ||
+      existing.description !== entry.description;
+    const timeChanged = existing.lastModified !== entry.lastModified;
+
+    indexData[idx] = { ...existing, ...entry };
+
+    if (metaChanged || timeChanged) {
+      op = metaChanged ? 'updated' : 'timestamp';
+    }
+  } else {
+    indexData.push(entry);
+    op = 'added';
+  }
+
+  const dedupMap = new Map();
+  indexData.forEach(e => {
+    const p = path.posix.normalize(e.path);
+    dedupMap.set(p, { ...dedupMap.get(p), ...e });
+  });
+  let deduped = Array.from(dedupMap.values());
+  deduped = sanitizeIndex(deduplicateEntries(deduped));
+  indexData.length = 0;
+  indexData.push(...deduped);
+
+  await persistIndex(indexData, repo, token, userId);
+  if (op === 'skipped') {
+    logDebug('[updateIndexEntry] skipped', normalized, 'no changes detected');
+  } else {
+    console.log('[updateIndexEntry]', op, normalized);
+  }
+  return entry;
+}
+
+function scanMemoryDir(dirPath) {
+  const results = [];
+
+  function walk(current) {
+    const items = fs.readdirSync(current);
+    items.forEach(item => {
+      const abs = path.join(current, item);
+      const stats = fs.statSync(abs);
+      if (stats.isDirectory()) {
+        walk(abs);
+      } else {
+        if (abs === indexFilename) return;
+        const rel = path.relative(path.join(__dirname, '..'), abs);
+        const meta = extractMeta(abs);
+        results.push({
+          path: rel,
+          type: categorizeMemoryFile(item),
+          ...meta,
+        });
+      }
+    });
+  }
+
+  if (fs.existsSync(dirPath)) walk(dirPath);
+  return results;
+}
+
+async function rebuildIndex(repo, token, userId) {
+  const paths = await scanMemoryFolderRecursively(repo, token);
+  const entries = [];
+  for (const rel of paths) {
+    const abs = path.join(path.join(__dirname, '..'), rel);
+    const meta = extractMeta(abs);
+    entries.push({
+      path: rel,
+      type: categorizeMemoryFile(path.basename(rel)),
+      title: meta.title,
+      description: meta.description,
+      lastModified: meta.lastModified,
+    });
+  }
+
+  const clean = sanitizeIndex(deduplicateEntries(entries));
+  await persistIndex(clean, repo, token, userId);
+  return clean;
+}
+
+async function updateIndexFileManually(newEntries, repo, token, userId) {
+  if (!Array.isArray(newEntries)) return [];
+  const results = [];
+  for (const entry of newEntries) {
+    if (!entry.path) continue;
+    try {
+      const updated = await updateIndexEntry(repo, token, entry, userId);
+      if (updated) results.push(updated);
+    } catch (e) {
+      console.error('[updateIndexFileManually]', e.message);
+    }
+  }
+  return results;
+}
+
+async function listMemoryFiles(repo, token, dirPath) {
+  const directory = dirPath.startsWith('memory') ? dirPath : path.join('memory', dirPath);
+  const fullPath = path.join(path.join(__dirname, '..'), directory);
+  if (!fs.existsSync(fullPath)) return [];
+
+  const results = [];
+  function walk(current) {
+    const entries = fs.readdirSync(current, { withFileTypes: true });
+    entries.forEach(entry => {
+      const abs = path.join(current, entry.name);
+      if (entry.isDirectory()) {
+        walk(abs);
+      } else if (entry.isFile()) {
+        const rel = path.relative(path.join(__dirname, '..'), abs).replace(/\\/g, '/');
+        if (rel.endsWith('index.json')) return;
+        if (/\.(md|txt|json|js|ts|jsx|tsx|html|css)$/i.test(entry.name)) {
+          results.push(rel);
+        }
+      }
+    });
+  }
+
+  walk(fullPath);
+  return results;
+}
+module.exports = {
+  updatePlan,
+  ensureContext,
+  listMemoryFiles,
+  writeFileSafe,
+  githubWriteFileSafe,
+  updateOrInsertJsonEntry,
+  updateIndexFile,
+  updateIndexFileManually,
+  scanMemoryFolderRecursively,
+  updateIndexEntry,
+  rebuildIndex,
+  updateIndexFromPath,
+  persistIndex,
+  fetchIndex,
+  loadIndex,
+  saveIndex,
+  fileExistsInRepo,
+  contextFilename,
+  planFilename,
+  indexFilename,
+  scanMemoryDir,
+  deduplicateEntries,
+  sanitizeIndex,
+  readIndexSafe,
+};
